# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Meno API
  version: 1.0.0
  description: >-
    RESTful API built with Express, MongoDB, and clean architecture principles. This API provides endpoints for
    authentication (email/password and Google OAuth), user management, and file storage with support for multiple
    storage providers (local and Google Cloud Storage).
  contact:
    name: API Support
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:6001
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authorization header using the Bearer scheme
  schemas:
...
    User:
      type: object
      required:
        - email
        - name
      properties:
        _id:
          type: string
          description: Auto-generated user ID
        email:
          type: string
          format: email
          description: User email (unique)
        name:
          type: string
          description: User full name
        password:
          type: string
          format: password
          description: User password (only for email/password auth)
        provider:
          type: string
          enum:
            - email
            - google
          description: Authentication provider
        googleId:
          type: string
          description: Google OAuth ID
        avatar:
          type: string
          description: Avatar file path
        status:
          type: string
          enum:
            - active
            - inactive
            - suspended
          description: User account status
        tier:
          type: string
          description: Reference to user's subscription tier
        tierStartDate:
          type: string
          format: date-time
          description: When the current tier started
        currentMonthUsage:
          type: object
          properties:
            duration:
              type: number
              description: Cached total duration used this month (seconds)
            lastReset:
              type: string
              format: date-time
              description: When the usage cache was last reset
            month:
              type: number
              description: Month this cache is for (1-12)
            year:
              type: number
              description: Year this cache is for
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
paths:
...
  /api/users:
    get:
      summary: Get all users
      description: Retrieve a list of all users in the system
      tags:
        - Users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
    post:
      summary: Create new user
      description: Create a new user with email and name
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address (must be unique)
                  example: john.doe@example.com
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  description: User's full name
                  example: John Doe
                status:
                  type: string
                  enum:
                    - active
                    - inactive
                    - suspended
                  description: User account status
                  example: active
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/me:
    get:
      summary: Get current authenticated user's profile
      description: Returns the profile of the currently authenticated user
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/{id}:
    get:
      summary: Get user by ID
      description: Retrieve a single user by their unique ID
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID (MongoDB ObjectId)
          example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update user
      description: Update an existing user's information
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID (MongoDB ObjectId)
          example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: john.updated@example.com
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  description: User's full name
                  example: John Updated
                status:
                  type: string
                  enum:
                    - active
                    - inactive
                    - suspended
                  description: User account status
                  example: active
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete user
      description: Delete a user from the system
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID (MongoDB ObjectId)
          example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User deleted successfully
        '400':
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/{id}/avatar:
    post:
      summary: Upload user avatar
      description: Upload an avatar image for a user (max 5MB)
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID (MongoDB ObjectId)
          example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - avatar
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Avatar image file (max 5MB)
      responses:
        '200':
          description: Avatar uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or file too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/me/usage:
    get:
      summary: Get current user's monthly usage statistics
      description: Retrieve usage statistics (total meeting duration) for the authenticated user for a specific month
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: year
          schema:
            type: integer
          description: Year (defaults to current year)
          example: 2025
        - in: query
          name: month
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Month (1-12, defaults to current month)
          example: 1
      responses:
        '200':
          description: Usage statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      userId:
                        type: string
                        example: 507f1f77bcf86cd799439011
                      year:
                        type: integer
                        example: 2025
                      month:
                        type: integer
                        example: 1
                      totalDurationSeconds:
                        type: number
                        example: 3600
                      totalDurationMinutes:
                        type: number
                        example: 60
                      totalDurationHours:
                        type: string
                        example: '1.00'
                      meetingCount:
                        type: integer
                        example: 5
                      period:
                        type: object
                        properties:
                          start:
                            type: string
                            format: date-time
                          end:
                            type: string
                            format: date-time
        '400':
          description: Invalid query parameters
        '401':
          description: Unauthorized
  /api/users/{userId}/meetings:
    get:
      summary: Get user's recent meetings
      description: Retrieve paginated list of meetings for a user across all their projects
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: User ID
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 5
        - in: query
          name: sort
          schema:
            type: string
            default: '-createdAt'
      responses:
        '200':
          description: User meetings retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: User not found
...